+++ lib/src/intl_suggestors/intl_migrator.dart
@@ -17,6 +17,7 @@ class ConstantStringMigrator extends GeneralizingAstVisitor
     with AstVisitingSuggestor {
   final IntlMessages _messages;
   final String _className;
+  // TODO: Move this and duplicate testing into IntlMessages.
   Set<String> names = {};
 
   ConstantStringMigrator(this._className, this._messages);
@@ -57,31 +58,17 @@ class ConstantStringMigrator extends GeneralizingAstVisitor
     // Make sure it's not private.
     var publicName =
         basicName.startsWith('_') ? basicName.substring(1) : basicName;
-    if (isUnique(publicName)) return publicName;
-    var contentBasedName =
-        toVariableName(stringContent(node.initializer as StringLiteral)!);
-    if (isUnique(contentBasedName)) return contentBasedName;
-    return appendANumberTo(publicName);
+    if (isUnique(publicName)) {
+      return publicName;
+    } else {
+      // Use a content-based name.
+      var contentBasedName =
+          toVariableName(stringContent(node.initializer as StringLiteral)!);
+      return contentBasedName;
+    }
   }
 
   bool isUnique(String name) => !names.contains(name);
-
-  /// Make a name that's unique within the constant methods in the Intl file.
-  ///
-  /// Do this by appending a number. It recursively tries each number in turn, which
-  /// would be pretty inefficient for large numbers, but if we have even dozens of duplicate
-  /// names then we need to change our approach more than just making unique numbers faster.
-  String appendANumberTo(String name, [int counter = 1]) {
-    // TODO: This is only unique within constants. It would be better to move this
-    // into a class representing the output.
-    if (!names.contains(name)) {
-      return name;
-    } else if (!names.contains('$name$counter')) {
-      return '$name$counter';
-    } else {
-      return appendANumberTo(name, counter + 1);
-    }
-  }
 }
 
 class IntlMigrator extends ComponentUsageMigrator {